BAU: Make expired subheading verification less brittle
When someone inputs an ambigous short code for a subheading that has
multiple subheadings with the same short code in its tree we'll rely
on the natural order in the database for fetching the either of the
relevant subheadings

This creates a non-deterministic situation whereby the subheading PLS
might change depending on the ordering.

Technically we can't say either is right since the short code is
ambigous so we just accept that any are right.