BAU: Fixes broken smoke tests
To support multiple tabs simultaneously, we have to make the storage of
goods nomenclature codes stateless (in the URL path as opposed to in the
session of the requests).

This means that on several pages with back links to goods nomenclature
and on forms, we have links rendered that include the
goods_nomenclature_code in the path and as hidden fields to
PATCH/POST/PUT in form submits.